# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QLGD API",
        "description": "Tài liệu OpenAPI cho hệ thống Quản lý Giảng dạy.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Máy chủ API"
        }
    ],
    "paths": {
...
        "/api/lecturer/notifications": {
            "get": {
                "tags": [
                    "Lecturer - Thông báo"
                ],
                "summary": "Danh sách thông báo của giảng viên",
                "operationId": "lecturerNotificationsIndex",
                "parameters": [
                    {
                        "name": "is_read",
                        "in": "query",
                        "description": "Lọc theo trạng thái đã đọc (true/false)",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách phân trang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NotificationResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturer/notifications/{id}": {
            "get": {
                "tags": [
                    "Lecturer - Thông báo"
                ],
                "summary": "Xem chi tiết thông báo",
                "operationId": "lecturerNotificationsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 501
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông báo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lecturer - Thông báo"
                ],
                "summary": "Xóa thông báo",
                "operationId": "lecturerNotificationsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 501
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Đã xóa"
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturer/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Lecturer - Thông báo"
                ],
                "summary": "Đánh dấu đã đọc thông báo",
                "operationId": "lecturerNotificationsMarkRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 501
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đánh dấu đã đọc thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Thông tin xác thực không hợp lệ."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
            "NotificationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 501
                    },
                    "type": {
                        "type": "string",
                        "example": "LEAVE_RESPONSE"
                    },
                    "title": {
                        "type": "string",
                        "example": "Đơn nghỉ đã được duyệt"
                    },
                    "body": {
                        "type": "string",
                        "example": "Phòng Đào tạo đã duyệt đơn nghỉ ngày 21/10."
                    },
                    "status": {
                        "type": "string",
                        "example": "UNREAD"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-20T14:00:00+07:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter the access token returned after login (omit the Bearer prefix)."
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}